chore: move Link classes to gui/link module
