+++ src/network_puzzles/gui/base.kv
@@ -12,7 +12,6 @@
 <FiltersCheckBox@ThemedCheckBox>:
     size_hint_x: 0.15
 
-
 <Packet>
     canvas:
         Color:
