+++ src/network_puzzles/gui/networkpuzzles.kv
@@ -1,12 +1,14 @@
 #:kivy 2.0
 #:import _ network_puzzles._
 #:import session network_puzzles.session
+# Order of "include" directives seems to matter.
+#:include network_puzzles/gui/popups.kv
 #:include network_puzzles/gui/base.kv
 #:include network_puzzles/gui/buttons.kv
 #:include network_puzzles/gui/device.kv
 #:include network_puzzles/gui/labels.kv
 #:include network_puzzles/gui/layouts.kv
-#:include network_puzzles/gui/popups.kv
+#:include network_puzzles/gui/link.kv
 
 
 ThemedBoxLayout:
@@ -102,3 +104,82 @@ ThemedBoxLayout:
         default_size_hint: 1, None
         size_hint_y: None
         height: self.minimum_height
+
+<PuzzleChooserPopup>:
+    size_hint: (1/3, 3/4)
+    title: _("Load Puzzle")
+    ThemedBoxLayout:
+        padding: dp(0)
+        spacing: dp(0)
+        canvas.before:
+            Color:
+                rgba: app.theme.bg2
+            Rectangle:
+                pos: self.pos
+                size: self.size
+        orientation: 'vertical'
+        ThemedBoxLayout:  # top section
+            padding: dp(0)
+            spacing: dp(0)
+            size_hint_y: 0.9
+            ThemedBoxLayout:
+                orientation: 'vertical'
+                CheckBoxLabel:
+                    text: _("Filter the puzzles by level:")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level0'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 0")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level1'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 1")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level2'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 2")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level3'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 3")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level4'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 4")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level5'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 5")
+                ThemedBoxLayout:
+                    FiltersCheckBox:
+                        name: 'Level6'
+                    CheckBoxLabel:
+                        size_hint_x: 0.85
+                        text: _("Level 6")
+            VSeparator:
+            PuzzlesRecView:
+                id: puzzles_view
+        HSeparator:
+        ThemedBoxLayout:  # bottom buttons
+            padding: dp(5)
+            spacing: dp(5)
+            size_hint_y: 0.1
+            size_hint_max_y: app.BUTTON_MAX_H
+            PopupButton:
+                text: _("Cancel")
+                on_press: root.on_cancel()
+            PopupButton:
+                text: _("Load")
+                on_press: root.on_load()
