+++ src/network_puzzles/gui/popups.kv
@@ -2,86 +2,6 @@
     title_size: dp(app.BUTTON_FONT_SIZE)
     size_hint: (1/2, 1/2)
 
-<PuzzleChooserPopup>:
-    size_hint: (1/3, 3/4)
-    title: _("Load Puzzle")
-    ThemedBoxLayout:
-        padding: dp(0)
-        spacing: dp(0)
-        canvas.before:
-            Color:
-                rgba: app.theme.bg2
-            Rectangle:
-                pos: self.pos
-                size: self.size
-        orientation: 'vertical'
-        ThemedBoxLayout:  # top section
-            padding: dp(0)
-            spacing: dp(0)
-            size_hint_y: 0.9
-            ThemedBoxLayout:
-                orientation: 'vertical'
-                CheckBoxLabel:
-                    text: _("Filter the puzzles by level:")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level0'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 0")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level1'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 1")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level2'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 2")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level3'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 3")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level4'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 4")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level5'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 5")
-                ThemedBoxLayout:
-                    FiltersCheckBox:
-                        name: 'Level6'
-                    CheckBoxLabel:
-                        size_hint_x: 0.85
-                        text: _("Level 6")
-            VSeparator:
-            PuzzlesRecView:
-                id: puzzles_view
-        HSeparator:
-        ThemedBoxLayout:  # bottom buttons
-            padding: dp(5)
-            spacing: dp(5)
-            size_hint_y: 0.1
-            size_hint_max_y: app.BUTTON_MAX_H
-            PopupButton:
-                text: _("Cancel")
-                on_press: root.on_cancel()
-            PopupButton:
-                text: _("Load")
-                on_press: root.on_load()
-
-
 <ExceptionPopup>:
     title: _("Error")
     ThemedBoxLayout:
@@ -109,36 +29,3 @@
             size_hint_y: 0.2
             text: _("Quit")
             on_release: app.stop()
-
-
-<CommandsPopup>
-    size_hint_x: 0.25
-
-<LinkPopup>
-    size_hint: (0.25, 0.33)
-    ThemedBoxLayout:
-        orientation: 'vertical'
-        #size_hint_y: None
-        PopupButton:
-            text: _("Delete")
-            on_press: root.delete()
-        PopupButton:
-            text: _("Edit")
-            on_press: root.edit()
-
-<PingHostPopup>:
-    size_hint: (0.25, 0.25)
-    ThemedBoxLayout:
-        orientation: 'vertical'
-        TextInput:
-            id: host
-            multiline: False
-            hint_text: _("hostname or IP")
-            on_text_validate: root.on_okay(root.ids.host.text)
-        ThemedBoxLayout:
-            PopupButton:
-                text: _("Cancel")
-                on_press: root.dismiss()
-            PopupButton:
-                text: _("Okay")
-                on_press: root.on_okay(root.ids.host.text)
